created events page to display all events, updated routes and controller to reflect changes. updated time and date format in event seeder and event details page. Updated flag aspect in cards component and updated styling. added events link to navbar

+++ resources/js/Pages/EventDetails/EventDetails.jsx
@@ -19,7 +19,7 @@ const EventDetails = ({ event }) => {
         "en-US",
         {
             day: "numeric",
-            month: "long",
+            month: "numeric",
             year: "numeric",
         }
     );
